updated comments
